Adição de client API e de envio de informações para o cliente

+++ Data/Scripts/AdvancedStatsAndEffects-CoreApi/PlayerData.cs
@@ -28,7 +28,7 @@ namespace AdvancedStatsAndEffects
             [XmlElement]
             public string Name { get; set; }
             [XmlElement]
-            public int Value { get; set; }
+            public long Value { get; set; }
 
         }
